% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSciCoverFun.R
\name{BSciCoverfun}
\alias{BSciCoverfun}
\title{Compute CI for bootstrap-calibrated methods}
\usage{
BSciCoverfun(beta, TB = NULL, G = NULL, alpha = 0.95)
}
\arguments{
\item{beta}{estimated betas}

\item{TB}{correction term}

\item{G}{indices of subgroups}

\item{alpha:}{confidence level}
}
\value{
\item{LowerBound}{Lower bound of the estimates}
\item{UpperBound}{Upper bound of the estimates}
\item{betaMax}{debiased maximum beta estimate}
}
\description{
Compute CI for bootstrap-calibrated methods
}
