% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSSplitLasso.R
\name{BSSplitLasso}
\alias{BSSplitLasso}
\title{Bootstrap-calibrated R-split method}
\usage{
BSSplitLasso(
  y,
  x,
  r = NULL,
  G = NULL,
  B = NULL,
  BB = NULL,
  alpha = 0.95,
  splitRatio = 0.6,
  fold = 2
)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}

\item{r}{tuning parameter}

\item{G}{subgroup indicator}

\item{B}{bootstrap number}

\item{BB}{split number}

\item{alpha}{level  ## change other places}

\item{splitRatio}{split ratio}

\item{fold}{cross validation fold}
}
\value{
\item{LowerBound}{lower confidence bound}
\item{UpperBound}{upper confidence bound}
\item{op}{optimal tuning}
}
\description{
Bootstrap-calibrated R-split method
}
