% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvDesparse.R
\name{cvDesparse}
\alias{cvDesparse}
\title{Select the optimal tuning for bootstrap-calibrated desparsified Lasso}
\usage{
cvDesparse(y, x, r = NULL, G = NULL, B = NULL, fold = 3)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}

\item{r}{candidate tuning parameters}

\item{G}{indices of subgroups}

\item{B}{bootstrap repetitions}

\item{fold}{number of folds in cross-validation}
}
\value{
\item{op}{optimal tuning parameter}
}
\description{
Select the optimal tuning for bootstrap-calibrated desparsified Lasso
}
