% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvSplit.R
\name{cvSplit}
\alias{cvSplit}
\title{Select the optimal tuning for bootstrap-calibrated R-Split through cross-validation}
\usage{
cvSplit(y, x, r = NULL, G = NULL, B = NULL, BB = NULL, ratio = NULL, fold = 2)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}

\item{r}{candidate tuning parameter}

\item{G}{subgroup indicator}

\item{B}{bootstrap iterations}

\item{BB}{split number}

\item{ratio}{ratio of data splitting}

\item{fold}{number of folds in cross-validation}
}
\value{
\item{op}{index of the optimal tuning}
}
\description{
Select the optimal tuning for bootstrap-calibrated R-Split through cross-validation
}
