% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFvarestbiascorr.R
\name{IFvarestbiascorr}
\alias{IFvarestbiascorr}
\title{variance calculation}
\usage{
IFvarestbiascorr(Ycount, alphaEst, n = NULL, splitSize = NULL)
}
\arguments{
\item{Ycount}{Y}

\item{alphaEst}{estimated values}

\item{n}{sample size}

\item{splitSize}{size of each split}
}
\value{
mean squared error
}
\description{
variance calculation
}
