% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zmatrix.R
\name{Zmatrix}
\alias{Zmatrix}
\title{Generate the nodewise Lasso matrix used in desparsified Lasso}
\usage{
Zmatrix(x, G = NULL)
}
\arguments{
\item{x}{nodewise confounder matrix}

\item{G}{indices of subgroups}
}
\value{
\item{Z}{nodewise Lasso matrix}
}
\description{
Generate the nodewise Lasso matrix used in desparsified Lasso
}
