% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSciCoverFun.R
\name{BSciCoverfun}
\alias{BSciCoverfun}
\title{Compute CI for bootstrap-calibrated methods}
\usage{
BSciCoverfun(beta, TB = NULL, G = NULL, alpha = 0.95)
}
\arguments{
\item{beta}{estimated betas}

\item{TB}{recalibrated bootstrap statistics}

\item{G}{indices of subgroups}

\item{alpha}{confidence level}
}
\value{
\item{LowerBound}{Lower confidence bound}
\item{UpperBound}{Upper confidence bound}
\item{betaMax}{debiased maximum beta estimate}
}
\description{
Compute CI for bootstrap-calibrated methods
}
