% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDesparseLasso.R
\name{BSDesparseLasso}
\alias{BSDesparseLasso}
\title{Bootstrap-calibrated Desparsified Lasso}
\usage{
BSDesparseLasso(y, x, r = NULL, G = NULL, B = NULL, alpha = 0.95, fold = 3)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}

\item{r}{tuning parameter}

\item{G}{subgroup indicator}

\item{B}{bootstrap iterations}

\item{alpha}{level of CI}
}
\value{
\item{LowerBound}{lower confidence bound}
\item{UpperBound}{upper confidence bound}
\item{op}{optimal tuning}
}
\description{
Bootstrap-calibrated Desparsified Lasso
}
