% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSSplitLasso.R
\name{BSSplitLasso}
\alias{BSSplitLasso}
\title{Bootstrap-calibrated R-split method}
\usage{
BSSplitLasso(
  y,
  x,
  r = NULL,
  G = NULL,
  B = NULL,
  BB = NULL,
  alpha = 0.95,
  splitRatio = 0.6,
  fold = 2
)
}
\arguments{
\item{y}{response}

\item{x}{design matrix}

\item{r}{tuning parameter}

\item{G}{subgroup indicator}

\item{B}{bootstrap number}

\item{BB}{split number}

\item{alpha}{level  ## change other places}

\item{splitRatio}{split ratio}

\item{fold}{cross validation fold}
}
\value{
\item{LowerBound}{lower confidence bound}
\item{UpperBound}{upper confidence bound}
\item{betaMax}{bias-reduced maximum beta estimate}
\item{betaEst}{debiased beta estimate for each subgroup}
\item{op}{optimal tuning}
}
\description{
This method first obtains the estimate of \eqn{\beta} via repetitive splitting
procedure (R-Split) through BB iterations. Then it calculates the calibration term
\eqn{\tilde{b}_{max} = (1-n^{r-0.5})(\tilde{\beta}_{max}-\tilde{\beta}_{j})}. Through B iterations,
it recalibrates the bootstrap statistic \eqn{T_b}. The bias-reduced estimate
is computed as: \eqn{\tilde{b}_{max}-\frac{1}{B}\sum_{b=1}^B T_b}.
}
