% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaMatNew.R
\name{sigmaMatNew}
\alias{sigmaMatNew}
\title{Generate covariance matrix}
\usage{
sigmaMatNew(p, type = NULL)
}
\arguments{
\item{p}{dimension of confounders}

\item{type}{type of matrix}
}
\value{
\item{Sigma}{A covariance matrix}
}
\description{
This function generates different types of covariance matrices
}
